refactor(profiling): extract stalk walking code
